Refactor node code into node.go
